pyproject.toml
[project]
name = "grumpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-allauth[socialaccount]>=65.12.0",
    "django-bootstrap-datepicker-plus>=5.0.5",
    "django-colorfield>=0.14.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.0.0",
    "django-environ>=0.12.0",
    "django-simple-captcha>=0.6.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "tzdata>=2025.2",
    "whitenoise[brotli]>=6.11.0",
    "zxcvbn>=4.5.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "factory-boy>=3.3.3",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-sugar>=1.1.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--no-migrations"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
testpaths = [
    "config",
    "grumpy",
]
